+++ cellprofiler/gui/view_workspace/control/_mask.py
@@ -1,5 +1,5 @@
 import cellprofiler.gui
-from cellprofiler.gui.view_workspace._frame import bind_data_class
+from cellprofiler.gui.view_workspace import bind_data_class
 from cellprofiler.gui.view_workspace.control._control import Control
